# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample service for managing drivers space",
        "title": "DriversSpace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Selmi hamza",
            "email": "hamza@craftfoundry.tech"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "api-company.herokuapp.com",
    "basePath": "/",
    "paths": {
...
        "/saascompanyAndoffice/{driver_id}": {
            "get": {
                "description": "retrieve a saascompany and saasOffice by driverID in database with the ID in request parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "saas_company_to_driver_relations saas_office_to_driver_relation"
                ],
                "summary": "retrieve a saascompany and saasOffice by driverID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "enter driver_id:",
                        "name": "driver_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CompanyAndOffice"
                        }
                    }
                }
            }
        },
...
        "/{saas_company}/offices": {
            "get": {
                "description": "retrieve a list of offices  in database  by saas company name in request parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "saas_companies saas_offices"
                ],
                "summary": "retrieve a list of offices by saas company name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "enter saas_company:",
                        "name": "saas_company",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.SaasOffice"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Models.SaasOffice": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "saasCompanyID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "controllers.CompanyAndOffice": {
            "type": "object",
            "properties": {
                "saasCompanyID": {
                    "type": "integer"
                },
                "saasOfficeID": {
                    "type": "integer"
                }
            }
        }
    }
}